refactor: 테스트 돕는 패키지명 변경
