+++ app/src/test/java/com/techcourse/service/helper/MockUserHistoryDao.java
@@ -1,4 +1,4 @@
-package com.techcourse.service.support;
+package com.techcourse.service.helper;
 
 import com.interface21.dao.DataAccessException;
 import com.interface21.jdbc.core.JdbcTemplate;
